Restructure the routing system using plain callbacks
